requirements.txt
icalevents==0.1.29
numpy==2.2.0
vobject==0.9.9
lxml==5.3.0
pillow==11.0.0
requests==2.32.3
spidev==3.6
gpiozero==2.0.1
lgpio==0.2.2.0
google-api-python-client==2.156.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
